+++ tapis-systemslib/src/main/java/edu/utexas/tacc/tapis/systems/model/PatchSystem.java
@@ -28,6 +28,7 @@ public final class PatchSystem
   private final String dtnMountPoint;
   private final String dtnMountSourcePath;
   private final Boolean canRunBatch;
+  private final String mpiCmd;
   private final List<JobRuntime> jobRuntimes;
   private final String jobWorkingDir;
   private final List<KeyValuePair> jobEnvVariables;
@@ -54,7 +55,7 @@ public final class PatchSystem
                      Integer port1, Boolean useProxy1, String proxyHost1, Integer proxyPort1,
                      String dtnSystemId1, String dtnMountPoint1, String dtnMountSourcePath1,
                      List<JobRuntime> jobRuntimes1, String jobWorkingDir1, List<KeyValuePair> jobEnvVariables1,
-                     Integer jobMaxJobs1, Integer jobMaxJobsPerUser1, Boolean canRunBatch1,
+                     Integer jobMaxJobs1, Integer jobMaxJobsPerUser1, Boolean canRunBatch1, String mpiCmd1,
                      SchedulerType batchScheduler1, List<LogicalQueue> batchLogicalQueues1,
                      String batchDefaultLogicalQueue1, String batchSchedulerProfile1,
                      List<Capability> jobCapabilities1, String[] tags1, Object notes1, String importRefId1)
@@ -71,6 +72,7 @@ public final class PatchSystem
     dtnMountPoint = dtnMountPoint1;
     dtnMountSourcePath = dtnMountSourcePath1;
     canRunBatch = canRunBatch1;
+    mpiCmd = mpiCmd1;
     jobRuntimes = (jobRuntimes1 == null) ? null : new ArrayList<>(jobRuntimes1);
     jobWorkingDir = jobWorkingDir1;
     jobEnvVariables = (jobEnvVariables1 == null) ? null : new ArrayList<>(jobEnvVariables1);
@@ -126,6 +128,8 @@ public final class PatchSystem
   public Integer getJobMaxJobsPerUser() { return jobMaxJobsPerUser; }
 
   public Boolean getCanRunBatch() { return canRunBatch; }
+
+  public String getMpiCmd() { return mpiCmd; }
 
   public SchedulerType getBatchScheduler() { return batchScheduler; }