+++ tapis-systemslib/src/test/java/edu/utexas/tacc/tapis/systems/IntegrationUtils.java
@@ -123,6 +123,9 @@ public final class IntegrationUtils
   public static final boolean canRunBatchTrue = true;
   public static final boolean canRunBatchFalse = false;
   public static final Boolean canRunBatchNull = null;
+  public static final String mpiCmd1 = "mpirun1";
+  public static final String mpiCmd2 = "mpirun2";
+  public static final String mpiCmdNull = null;
   public static final int jobMaxJobs1 = 1;
   public static final int jobMaxJobs2 = 2;
   public static final Integer jobMaxJobsNull = null;
@@ -255,7 +258,7 @@ public final class IntegrationUtils
             prot1.getPort(), prot1.isUseProxy(), prot1.getProxyHost(), prot1.getProxyPort(),
             dtnSystemIdNull, dtnMountPointNull, dtnMountSourcePathNull, isDtnTrue,
             canExecFalse, jobRuntimesNull, jobWorkingDirNull, jobEnvVariablesNull, jobMaxJobs1, jobMaxJobsPerUser1,
-            canRunBatchFalse, batchSchedulerNull, logicalQueueListNull, queueNameNull, batchSchedulerProfileNull,
+            canRunBatchFalse, mpiCmdNull, batchSchedulerNull, logicalQueueListNull, queueNameNull, batchSchedulerProfileNull,
             capListNull, tags1, notes1, importRefId1, uuidNull, isDeletedFalse, createdNull, updatedNull);
   }
 
@@ -274,7 +277,7 @@ public final class IntegrationUtils
             prot2.getPort(), prot2.isUseProxy(), prot2.getProxyHost(), prot2.getProxyPort(),
             dtnSystemIdNull, dtnMountPointNull, dtnMountSourcePathNull, isDtnTrue,
             canExecFalse, jobRuntimesNull, jobWorkingDirNull, jobEnvVariablesNull, jobMaxJobs2, jobMaxJobsPerUser2,
-            canRunBatchFalse, batchSchedulerNull, logicalQueueListNull, queueNameNull, batchSchedulerProfileNull,
+            canRunBatchFalse, mpiCmdNull, batchSchedulerNull, logicalQueueListNull, queueNameNull, batchSchedulerProfileNull,
             capListNull, tags2, notes2, importRefId2, uuidNull, isDeletedFalse, createdNull, updatedNull);
   }
 
@@ -299,14 +302,14 @@ public final class IntegrationUtils
             prot1.getPort(), prot1.isUseProxy(), prot1.getProxyHost(), prot1.getProxyPort(),
             dtnSystemIdNull, dtnMountPointNull, dtnMountSourcePathNull, isDtnTrue,
             canExecFalse, jobRuntimesNull, jobWorkingDirNull, jobEnvVariablesNull, jobMaxJobs1, jobMaxJobsPerUser1,
-            canRunBatchFalse, batchSchedulerNull, logicalQueueListNull, queueNameNull, batchSchedulerProfileNull,
+            canRunBatchFalse, mpiCmdNull, batchSchedulerNull, logicalQueueListNull, queueNameNull, batchSchedulerProfileNull,
             capListNull, tags1, notes1, importRefId1, uuidNull, isDeletedFalse, createdNull, updatedNull);
     dtnSystem2 = new TSystem(-1, tenantName, dtnSystemName2, "DTN System2 for tests", TSystem.SystemType.LINUX, owner1,
             dtnSystemValidHostname, isEnabledTrue,"effUserDtn2", prot2.getAuthnMethod(), "bucketDtn2", "/root/dtn2",
             prot2.getPort(), prot2.isUseProxy(), prot2.getProxyHost(), prot2.getProxyPort(),
             dtnSystemIdNull, dtnMountPointNull, dtnMountSourcePathNull, isDtnTrue,
             canExecFalse, jobRuntimesNull, jobWorkingDirNull, jobEnvVariablesNull, jobMaxJobs2, jobMaxJobsPerUser2,
-            canRunBatchFalse, batchSchedulerNull, logicalQueueListNull, queueNameNull, batchSchedulerProfileNull,
+            canRunBatchFalse, mpiCmdNull, batchSchedulerNull, logicalQueueListNull, queueNameNull, batchSchedulerProfileNull,
             capListNull, tags2, notes2, importRefId2, uuidNull, isDeletedFalse, createdNull, updatedNull);
     for (int i = 0; i < n; i++)
     {
@@ -321,7 +324,7 @@ public final class IntegrationUtils
               prot1.getPort(), prot1.isUseProxy(), prot1.getProxyHost(), prot1.getProxyPort(),
               dtnSystem1.getId(), dtnMountPoint1, dtnMountSourcePath1, isDtnFalse,
               canExecTrue, jobRuntimes1, "jobWorkDir"+suffix, jobEnvVariables1, jobMaxJobs1, jobMaxJobsPerUser1,
-              canRunBatchTrue, batchScheduler1, logicalQueueList1, queueA1.getName(), batchSchedulerProfile1,
+              canRunBatchTrue, mpiCmd1, batchScheduler1, logicalQueueList1, queueA1.getName(), batchSchedulerProfile1,
               capList1, tags1, notes1, importRefId1, uuidNull, isDeletedFalse, createdNull, updatedNull);
       systems[i].setJobRuntimes(jobRuntimes1);
       systems[i].setBatchLogicalQueues(logicalQueueList1);
@@ -345,7 +348,7 @@ public final class IntegrationUtils
               hostMinimalId, isEnabledTrue, effectiveUserIdNull, tSys.getDefaultAuthnMethod(), null, rootDir1,
               prot1.getPort(), prot1.isUseProxy(), null, prot1.getProxyPort(), null, null, null, isDtnFalse,
               canExecFalse, jobRuntimesNull, null, null, jobMaxJobs1, jobMaxJobsPerUser1,
-              canRunBatchFalse, batchSchedulerNull, logicalQueueListNull, batchDefaultLogicalQueueNull,
+              canRunBatchFalse, mpiCmdNull, batchSchedulerNull, logicalQueueListNull, batchDefaultLogicalQueueNull,
               batchSchedulerProfileNull, capListNull,
               tagsNull, notesNull, importRefIdNull, uuidNull, isDeletedFalse, createdNull, updatedNull);
     }
@@ -355,7 +358,7 @@ public final class IntegrationUtils
               hostMinimalId, isEnabledTrue, effectiveUserIdNull, tSys.getDefaultAuthnMethod(), null, rootDir1,
               prot1.getPort(), prot1.isUseProxy(), null, prot1.getProxyPort(), null, null, null, isDtnFalse,
               canExecFalse, jobRuntimesNull, null, null, jobMaxJobs1, jobMaxJobsPerUser1,
-              canRunBatchFalse, batchSchedulerNull, logicalQueueListNull, batchDefaultLogicalQueueNull,
+              canRunBatchFalse, mpiCmdNull, batchSchedulerNull, logicalQueueListNull, batchDefaultLogicalQueueNull,
               batchSchedulerProfileNull, capListNull,
               tagsNull, notesNull, importRefIdNull, uuidNull, isDeletedFalse, null, null);
     }
@@ -374,8 +377,9 @@ public final class IntegrationUtils
                        prot2.getAuthnMethod(), system.getBucketName(), system.getRootDir(), prot2.getPort(),
                        prot2.isUseProxy(), prot2.getProxyHost(), prot2.getProxyPort(),
                        sysNamePrefix+key+dtnSystemId2, dtnMountPoint2, dtnMountSourcePath2, system.isDtn(),
-                       system.getCanExec(), jobRuntimes2, jobWorkingDir2, jobEnvVariables2, jobMaxJobs2, jobMaxJobsPerUser2,
-            canRunBatchTrue, batchScheduler2, logicalQueueList2, batchDefaultLogicalQueue2, batchSchedulerProfile2,
+                       system.getCanExec(), jobRuntimes2, jobWorkingDir2, jobEnvVariables2, jobMaxJobs2,
+                       jobMaxJobsPerUser2, canRunBatchTrue, mpiCmd2, batchScheduler2, logicalQueueList2,
+                       batchDefaultLogicalQueue2, batchSchedulerProfile2,
                        capList2, tags2, notes2, importRefId2, null, false, null, null);
     putSys.setBatchLogicalQueues(logicalQueueList2);
     putSys.setJobRuntimes(jobRuntimes2);
@@ -394,7 +398,7 @@ public final class IntegrationUtils
     return new PatchSystem(description2, hostname2, effectiveUserId2,
             prot2.getAuthnMethod(), prot2.getPort(), prot2.isUseProxy(), prot2.getProxyHost(), prot2.getProxyPort(),
             sysNamePrefix+key+dtnSystemId2, dtnMountPoint2, dtnMountSourcePath2, jobRuntimes2, jobWorkingDir2,
-            jobEnvVariables2, jobMaxJobs2, jobMaxJobsPerUser2, canRunBatchTrue, batchScheduler2,
+            jobEnvVariables2, jobMaxJobs2, jobMaxJobsPerUser2, canRunBatchTrue, mpiCmd2, batchScheduler2,
             logicalQueueList2, batchDefaultLogicalQueue2, batchSchedulerProfile2, capList2, tags2, notes2, importRefId2);
   }
 
@@ -407,7 +411,7 @@ public final class IntegrationUtils
     return new PatchSystem(description2, hostnameNull, effectiveUserIdNull,
             prot2.getAuthnMethod(), portNull, userProxyNull, proxyHostNull, proxyPortNull,
             dtnSystemIdNull, dtnMountPoint2, dtnMountSourcePathNull, jobRuntimes2, jobWorkingDirNull, jobEnvVariablesNull,
-            jobMaxJobsNull, jobMaxJobsPerUser2, canRunBatchNull, batchSchedulerNull, logicalQueueListNull,
+            jobMaxJobsNull, jobMaxJobsPerUser2, canRunBatchNull, mpiCmd2, batchSchedulerNull, logicalQueueListNull,
             batchDefaultLogicalQueueNull, batchSchedulerProfileNull, capListNull, tagsNull, notesNull, importRefIdNull);
   }