+++ CHANGELOG.md
@@ -8,6 +8,17 @@ https://tapis.readthedocs.io/en/latest/technical/systems.html
 You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
+## 1.0.3 - 2021-12-15
+
+Incremental improvements and bug fixes.
+
+### New features:
+- Add mpiCmd attribute.
+
+### Bug fixes:
+- Fix bug in setting default for SchedulerProfile owner.
+
+
 ## 1.0.2 - 2021-11-12
 
 Incremental improvements and bug fixes.